wit/bindgen: move generator code into generate.go
